+++ src/main/java/com/leverx/javacourse/seller_rating_app/repository/ItemRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ItemRepository extends CrudRepository<Item, Integer> {
+public interface ItemRepository extends CrudRepository<Item, Long> {
     List<Item> findByGameTitle(String gameTitle);
 }
